+++ Backend/src/plugins/handlers/userHandlers.js
@@ -1,18 +1,35 @@
-const responseHelper = require('../helpers/responseHelper');
+const { getToken } = require('../helpers/authHelper');
+const {
+  response200Handler,
+  response201Handler,
+  response400Handler,
+  response400HandlerImage,
+  response401Handler,
+  response404Handler,
+} = require('../helpers/responseHelper');
 const {
   validateImageExtension,
   deleteSavedImage,
   saveImage,
 } = require('../helpers/saveImageHelper');
+const { adminChecker, userChecker } = require('../helpers/usersChecker');
 
 const getUser = async (request, h) => {
+  const { userId: id } = request.auth.credentials;
   const { prisma } = request.server.app;
-  const user = await prisma.user.findMany({
+
+  const requesterUser = await adminChecker(prisma, h, id, 'get');
+  if (requesterUser.error) {
+    return requesterUser.dataError;
+  }
+
+  const users = await prisma.user.findMany({
     where: {
       roleId: 2,
     },
   });
-  return responseHelper(h, 'success', 'Data berhasil didapatkan', 200, user);
+
+  return response200Handler(h, 'get', users);
 };
 
 const getUserById = async (request, h) => {
@@ -20,12 +37,7 @@ const getUserById = async (request, h) => {
   const { id } = request.params;
 
   if (!id) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal mendapatkan user. Mohon isi id user.',
-      400,
-    );
+    return response400Handler(h, 'get', 'user', 'id');
   }
 
   const user = await prisma.user.findUnique({
@@ -35,15 +47,10 @@ const getUserById = async (request, h) => {
   });
 
   if (user) {
-    return responseHelper(h, 'success', 'Data berhasil didapatkan', 200, user);
+    return response200Handler(h, 'get', user);
   }
 
-  return responseHelper(
-    h,
-    'failed',
-    'Gagal mendapatkan data user. Id tidak ditemukan',
-    404,
-  );
+  return response404Handler(h, 'get', 'user', 'Id');
 };
 
 const addUser = async (request, h) => {
@@ -53,21 +60,11 @@ const addUser = async (request, h) => {
   let dataImage;
 
   if (!name) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menambahkan user. Mohon isi nama user',
-      400,
-    );
+    return response400Handler(h, 'add', 'user', 'name');
   }
 
   if (!email) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menambahkan user. Mohon isi email user',
-      400,
-    );
+    return response400Handler(h, 'add', 'user', 'email');
   }
 
   const check = await prisma.user.findUnique({
@@ -77,12 +74,7 @@ const addUser = async (request, h) => {
   });
 
   if (check) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menambahkan user. Email user sudah ada',
-      400,
-    );
+    return response400Handler(h, 'add', 'user', null, 'Email user sudah ada');
   }
 
   if (validateImageExtension(image)) {
@@ -90,12 +82,7 @@ const addUser = async (request, h) => {
       dataImage = await saveImage(image, 'user');
     }
   } else {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menambahkan gambar. Mohon memberikan gambar dengan ekstensi [jpg, jpeg, atau png]',
-      400,
-    );
+    return response400HandlerImage(h);
   }
 
   const user = await prisma.user.create({
@@ -108,29 +95,25 @@ const addUser = async (request, h) => {
     },
   });
 
-  return responseHelper(h, 'success', 'User berhasil ditambahkan', 201, user);
+  return response201Handler(h, 'User', user);
 };
 
 const updateUser = async (request, h) => {
+  const { userId: id } = request.auth.credentials;
   const { prisma } = request.server.app;
   const { name, email, oldImage, newImage } = request.payload;
 
+  const requesterUser = await userChecker(prisma, h, id, 'update');
+  if (requesterUser.error) {
+    return requesterUser.dataError;
+  }
+
   if (!name) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal memperbarui user. Mohon isi nama user',
-      400,
-    );
+    return response400Handler(h, 'update', 'user', 'name');
   }
 
   if (!email) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal memperbarui user. Mohon isi email user',
-      400,
-    );
+    return response400Handler(h, 'update', 'user', 'email');
   }
 
   const userNow = await prisma.user.findUnique({
@@ -140,12 +123,11 @@ const updateUser = async (request, h) => {
   });
 
   if (!userNow) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal memperbarui user. Email tidak ditemukan',
-      404,
-    );
+    return response404Handler(h, 'update', 'user', 'Email');
+  }
+
+  if (requesterUser.data.roleId != 1 || userNow.id === requesterUser.data.id) {
+    return response401Handler(h, 'user request');
   }
 
   let image_large = userNow.image_large;
@@ -173,52 +155,79 @@ const updateUser = async (request, h) => {
     },
   });
 
-  return responseHelper(h, 'success', 'User berhasil diperbarui', 200, user);
+  return response200Handler(h, 'update', user);
 };
 
 const deleteUser = async (request, h) => {
+  const { userId: id } = request.auth.credentials;
+
   const { prisma } = request.server.app;
+
+  const requesterUser = await userChecker(prisma, h, id, 'delete');
+  if (requesterUser.error) {
+    return requesterUser.dataError;
+  }
+
   const { email } = request.payload;
 
   if (!email) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menghapus user. Mohon berikan email user',
-      400,
-    );
+    return response400Handler(h, 'delete', 'user', 'email');
   }
 
-  let user;
+  let user = await prisma.user.findUnique({ where: { email } });
 
-  try {
-    user = await prisma.user.delete({
-      where: {
-        email,
-      },
-    });
-  } catch (e) {
-    return responseHelper(
-      h,
-      'failed',
-      'Gagal menghapus user. Email tidak ditemukan',
-      404,
-    );
+  if (!user) {
+    return response404Handler(h, 'delete', 'user', 'Email');
   }
 
-  return responseHelper(
-    h,
-    'success',
-    'Berhasil menghapus data user.',
-    200,
-    user,
-  );
+  if (requesterUser.roleId != 1 || requesterUser.id !== user.id) {
+    return response401Handler(h, 'role');
+  }
+
+  user = await prisma.user.delete({
+    where: {
+      email,
+    },
+  });
+
+  return response200Handler(h, 'delete', user);
 };
 
 // For view image
 const getImage = async (request, h) => {
   const { name } = request.params;
   return h.file(`./uploads/user/${name}`);
+};
+
+const userLogin = async (request, h) => {
+  const { prisma } = request.server.app;
+  const { email, password } = request.payload;
+
+  if (!email) {
+    return response400Handler(h, 'get', 'user', 'email');
+  }
+
+  if (!password) {
+    return response400Handler(h, 'get', 'user', 'password');
+  }
+
+  const user = await prisma.user.findUnique({ where: { email } });
+
+  if (!user) {
+    return response404Handler(h, 'get', 'user', 'Email');
+  }
+
+  if (user.roleId != 2) {
+    return response401Handler(h, 'role');
+  }
+
+  if (user.password !== password) {
+    return response400Handler(h, 'get', 'admin', 'password dengan benar');
+  }
+
+  const token = getToken(user.id);
+
+  return response200Handler(h, 'get', token);
 };
 
 module.exports = {
@@ -228,4 +237,5 @@ module.exports = {
   deleteUser,
   addUser,
   getImage,
+  userLogin,
 };
